fix: use DOUBLE PRECISION in V27 migration for PostgreSQL

+++ backend/src/main/resources/db/migration/V27__org_slo_settings.sql
@@ -1,4 +1,4 @@
 ALTER TABLE organizations
-    ADD COLUMN slo_availability_target DOUBLE NOT NULL DEFAULT 0.995,
-    ADD COLUMN slo_latency_target DOUBLE NOT NULL DEFAULT 0.95,
+    ADD COLUMN slo_availability_target DOUBLE PRECISION NOT NULL DEFAULT 0.995,
+    ADD COLUMN slo_latency_target DOUBLE PRECISION NOT NULL DEFAULT 0.95,
     ADD COLUMN slo_latency_threshold_seconds INT NOT NULL DEFAULT 2;
